+++ packages/build-app-helpers/tests/fixtures/analyzeAuth/app.js
@@ -1,0 +1,5 @@
+import { runApp } from 'ice';
+
+runApp({
+  auth: {},
+});